% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{computeFeatures.geoLet}
\alias{computeFeatures.geoLet}
\title{computeFeatures.geoLet lavora su piu}
\usage{
computeFeatures.geoLet(obj.geoLet, ROIName, feature.family = c("stat",
  "morph", "glcm", "rlm", "szm"), filterPipeline = c(), px = "",
  py = "", threshold = 0.1, discretize = "", bin.size = "",
  bin.number = "", n_grey = 100)
}
\description{
Intiate an object of the class \code{geoLet}.This represents just the classname (f.extractor.sing.par)
}
