% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{f.extractor.pluri.LoG.par}
\alias{f.extractor.pluri.LoG.par}
\title{Function extractor for multiple filter parameter}
\usage{
f.extractor.pluri.LoG.par(path, ROIName, feature.family = c("stat",
  "morph", "glcm", "rlm", "szm"), interpolate = FALSE, px = "",
  py = "", threshold = 0.1, discretize = "", bin.size = 25,
  bin.number = 25, fileName = "tmp.f.extractor.pluri.par.RData",
  forceRecalculus = TRUE, from.sigma = from.sigma,
  to.sigma = to.sigma, def.by = def.by, strategy = "fixed")
}
\arguments{
\item{path}{The path to be browsed for folders containing DICOM images and Structure Set of patients}

\item{ROIName}{A \code{character} object containing the name(s) (as vector) of ROIs to be analyzed. If there is more than one ROI (multiple references) or no ROI is listed the analysis is stopped}

\item{feature.family}{The list of features to be extracted, by default all. The possible families are: statistical, grey level co-occurrence matrix, run-length matrix, size-zone matrix and fractal}

\item{interpolate}{A \code{logical} value, by default \code{FALSE}, for interpolating sigma values in filters using sigma for computation as LoG}

\item{px}{The pixel spacing in \emph{x} direction as string for interpolating sigma, by default empty.}

\item{py}{The pixel spacing in \emph{y} direction as string for interpolating sigma, by default empty.}

\item{threshold}{A \code{numeric} value setting the threshold for applying the interpolation, by default .001}

\item{discretize}{ROI discretization for feature extractions, by default \code{none}. The possible values are \code{fixed.bin.size} and \code{fixed.bin.number}.}

\item{bin.size}{Size of bin for ROI discretization for feature extractions, by default 25.}

\item{bin.number}{Number of bin for ROI discretization for feature extractions, by default 25.}

\item{fileName}{The Rdata file name saved locally during the running, by default 'tmp.f.extractor.sing.par.RData'.}

\item{forceRecalculus}{A \code{logical} value, by defaul \code{TRUE}, for forcing recalculus by deleting all the local Rdata previously created.}

\item{from.sigma}{The first sigma to be analyzed}

\item{to.sigma}{The last sigma to be analyzed}

\item{def.by}{The sigma step to be analyzed}

\item{strategy}{The strategy to be applied for the radiomics analysis, by default "fixed". For now no other strategy have been implemented.}
}
\description{
It extracts the feature values by applying the filter at multiple LoG sigma.
}
\examples{
\dontrun{
f.extractor.pluri.LoG.par(path = './RadiomicsFolder', ROIName = 'GTV', feature.family = c('stat', 'glcm'), forceRecalculus = T, from.sigma = 0.3, to.sigma = 2, def.by = 0.1)
}
}
