% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{f.extractor.sing.par}
\alias{f.extractor.sing.par}
\title{Function extractor for single filter parameter}
\usage{
f.extractor.sing.par(path, ROIName, feature.family = c("stat", "morph",
  "glcm", "rlm", "szm", "fractal"), filterPipeline = list(),
  interpolate = FALSE, px = "", py = "", threshold = 0.001,
  discretize = "", bin.size = 25, bin.number = 25,
  fileName = "tmp.f.extractor.sing.par.RData", forceRecalculus = TRUE)
}
\arguments{
\item{path}{The path to be browsed for folders containing DICOM images and Structure Set of patients}

\item{ROIName}{A \code{character} object containing the name(s) (as vector) of ROIs to be analyzed. If there is more than one ROI (multiple references) or no ROI is 
listed the analysis is stopped}

\item{feature.family}{The list of features to be extracted, by default all. The possible families are: statistical, grey level co-occurrence matrix, run-length matrix, size-zone matrix and fractal}

\item{filterPipeline}{A \code{list} contains the filter pipeline to be applied, by default empty, to images before feature extraction. The first list argument is "kernel.type" as character and the second argument is "sigma" as numeric.}

\item{interpolate}{A \code{logical} value, by default \code{FALSE}, for interpolating sigma values in filters using sigma for computation as LoG}

\item{px}{The pixel spacing in \emph{x} direction as string for interpolating sigma, by default empty.}

\item{py}{The pixel spacing in \emph{y} direction as string for interpolating sigma, by default empty.}

\item{threshold}{A \code{numeric} value setting the threshold for applying the interpolation, by default .001}

\item{discretize}{ROI discretization for feature extractions, by default \code{none}. The possible values are \code{fixed.bin.size} and \code{fixed.bin.number}.}

\item{bin.size}{Size of bin for ROI discretization for feature extractions, by default 25.}

\item{bin.number}{Number of bin for ROI discretization for feature extractions, by default 25.}

\item{fileName}{The Rdata file name saved locally during the running, by default 'tmp.f.extractor.sing.par.RData'.}

\item{forceRecalculus}{A \code{logical} value, by defaul \code{TRUE}, for forcing recalculus by deleting all the local Rdata previously created.}
}
\description{
It extracts the feature values by applying the filter at a single LoG sigma.
}
\details{
Using an object of class \code{ROImap} it is possible to list all the ROIs contained in an explored directory for the analysis. It is possible to analyze
different ROIs names referencing to the same structure in the different studies by putting al their names a a vector of \code{character} objects in the argument \code{ROIName}
}
\examples{
\dontrun{
# Create a filter pipeline list
filterPip <- list()
filterPip[[1]] <- list('kernel.type'='LoG', 'sigma' = 0.3)
f.extractor.sing.par(path = './RadiomicsFolder', ROIName = 'GTV', feature.family = c('stat', 'glcm'), filterPipeline = filterPip, forceRecalculus = T)
}
}
