% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fractalFeatures.R
\name{fractalFeatures}
\alias{fractalFeatures}
\title{Fractal features}
\usage{
fractalFeatures(imgObj, fstPerc = 0, lstPerc = 100, ThDown = 0,
  ThUp = 100, onepxwide = T)
}
\arguments{
\item{imgObj}{The imgObj get as input the GTV object extracted using getROIVoxels method.}

\item{fstPerc}{A \code{numeric} value for normalized the image. It is the first percentile.}

\item{lstPerc}{A \code{numeric} value for normalized the image. It is the last percentile.}

\item{ThDown}{A \code{numeric} value for masked image filter. It is the minimum value of the threshold}

\item{ThUp}{A \code{numeric} value for masked image filter. It is the maximum value of the threshold}
}
\description{
This function extracts fractal features using the box counting technique.
}
\examples{
\dontrun{
# Create an instance of geoLet object and load a case
obj <- geoLet()
obj$openDICOMFolder(pathToOpen = './patient1')

# get the ROI
GTV <- obj$getROIVoxels(Structure = 'GTV')

# extract the fractal feature
 Fractal <- fractalFeatures(imgObj = GTV, fstPerc = 1, lstPerc = 99, ThDown = 0, ThUp = 100)

}
}
