% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoNiftiFunction.R
\name{geoNIFTI}
\alias{geoNIFTI}
\title{class for loading and presenting NIFTI data}
\usage{
geoNIFTI()
}
\description{
Instantiate an object of the class \code{geoNIFTI}.This represents just the classname, 
              methods are exposed with the technique of 'closure'.
              In order to see manuals for the single mathods, consider the vignette or use the 
              available for the following wrapping functions:
              \itemize{
              \item \code{GLT.getImageVoxelCube( );} : to get the ImageVoxelCube stored into a geoNIFTI object
              \item \code{GLT.getROIVoxels( );} : to get the IMAGE Voxels geometrically located into a ROI, for a given geoNIFTI object
              }
              The original methods for the class geoLet can also be invocked using the same name without the previx 'GTL.', i.e.:
}
\examples{
\dontrun{
# first of all create an object obj
obj.geoNIFTI<-geoNIFTI()
tmp<-obj.geoNIFTI$getROIVoxels("/Users/davpas/Desktop/bis/BOOGLO","*metastatic*")
tmp2<- obj.geoNIFTI$getImageVoxelCube("/Users/davpas/Desktop/bis/BOOGLO")
matrixImageROI<-tmp$masked.images$voxelCube

}
}
