% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fractalFeatures.R
\name{NormerObj}
\alias{NormerObj}
\title{Funzione 3}
\usage{
NormerObj(imgObj, fstPerc = 0, lstPerc = 100)
}
\description{
NormerObj prende in ingresso un geoLet che ha già caricato l'immagine ed estratto la ROI
Questa funzione restituisce la stessa mButoVoxelList, ma con dentro i VoxelCube le immagini normalizzate
}
