% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{plot2D.Slice}
\alias{plot2D.Slice}
\title{function for plotting a series of slices containing (or not) ROI(s) to be shown}
\usage{
plot2D.Slice(path, ROIName, col)
}
\arguments{
\item{path}{The path of the series to be shown}

\item{ROIName}{A \code{character} object containing the name(s) of the ROI(s) to be shown. If empty all slices are shown in sequence.}

\item{col}{A \code{character} object listing the color(s) for showing the ROI(s). If empty ROI(s) will be plotted in red.}
}
\description{
This function has been designed for plotting DICOM slice(s) containing one or more ROIs.
}
