% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADscoutNIFTI.R
\name{RAD.scoutNIFTI}
\alias{RAD.scoutNIFTI}
\title{class for loading and presenting DICOM data}
\usage{
RAD.scoutNIFTI(path, regexp_roi)
}
\description{
Create a report using as input a path containing directories with patient,
              \itemize{
              \item \code{setPath(), getPath(), pixelSpacing;} : to load a DICOM series into an geoLet object
              }
}
\examples{
\dontrun{
# first of all create an object aa 
aa <- RAD.scoutNIFTI(path = '/Users/davpas/Desktop/bis', '*metastatic*')
aa$setPath("/Users/davpas/Desktop/moddicom/moddicom/R/")
aa$pixelSpacing()

}
}
