% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_fractalFeatures.R
\name{ThresholderObj}
\alias{ThresholderObj}
\title{Funzione 2}
\usage{
ThresholderObj(imgObj, ThDown = 0, ThUp = 100)
}
\description{
ThresholderObj prende in ingresso un geoLet che ha già caricato l'immagine ed  estratto la ROI
Questa funzione restituisce lo stesso Geolet, ma con dentro i VoxelCube le immagini binarizzate
}
