% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{writeLog}
\alias{writeLog}
\title{Write an external file}
\usage{
writeLog(stringa, file = path)
}
\description{
scrive quanto indicato come stringa nell'oggetto stringa
 su un file esterno nel path indicato nel LogFilePath
}
